server/requirements.txt
fastapi 
uvicorn 
motor 
redis
numpy 
pydantic-settings
python-multipart
aiohttp 
pillow
facenet-pytorch 
torch 
torchvision 
